gr-qtgui: Fix some Type Errors in GUI Widgets
QT GUI Compass, Led Indicator, Digital Number Control

fail with Type Error

+++ gr-qtgui/python/qtgui/digitalnumbercontrol.py
@@ -298,7 +298,7 @@ class DigitalNumberControl(QFrame):
         rect = QtCore.QRect(2, 2, size.width() - 4, size.height() - 4)
         painter.fillRect(rect, brush)
 
-        self.numberFont.setPixelSize(0.9 * size.height())
+        self.numberFont.setPixelSize(int(0.9 * size.height()))
         painter.setFont(self.numberFont)
         painter.setPen(self.fontColor)
         rect = event.rect()
